### 2026-05-18 #10584 @editorialbot
[PRE REVIEW]: ABCdeZ.jl: Simulation-based Bayesian inference and model evidence estimation in Julia
**Submitting author:** @mauricelanghinrichs (<a href="http://orcid.org/0000-0001-8529-5749">Maurice Langhinrichs</a>)
**Repository:** https://github.com/mauricelanghinrichs/ABCdeZ.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.6.0
