### 2026-08-28 #10584 @jbytecode
@dwannetts, @slwu89 -  Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Please note that this the second reply due to the loss of notications.
