### 2026-08-30 #11223 @dwannetts
My only comment would be to add a seed() to allow easier reproduction of the minimal working example.  I was able to run the example a few times to convince myself that reproduction of the figure was possible in general, but not guaranteed.
