+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#ABCdeZ.jl","page":"ABCdeZ.jl","title":"ABCdeZ.jl","text":"Approximate Bayesian Computation (ABC) with differential evolution (de) moves and model evidence (Z) estimates.\n\nABCdeZ.jl offers Bayesian parameter estimation and model comparison/selection for inference problems with an intractable likelihood. Models only need to be simulated (instead of calculating the likelihood). In this documentation you will find everything to get started.\n\n\n\nABCdeZ.jl was developed @TSB by Maurice Langhinrichs and Nils Becker. This work is based on many people's previous achievements, particular some part of the code base was adapted from KissABC.jl [1]; please find a complete list of references below.","category":"section"},{"location":"#Brief-introduction","page":"ABCdeZ.jl","title":"Brief introduction","text":"Bayesian inference allows to update a parameter prior p(θ  M) in the light of new data D to obtain posterior knowledge as\n\np(θ  D M) = fracp(D θ M)  p(θ  M)Z\n\nwhere θ are the parameters of model M, p(D θ M) the likelihood and Z = p(D  M) is the model evidence (also known as marginal likelihood). The model evidence is the normalising integral \n\nZ equiv p(D  M) = int  p(D θ M)  p(θ  M)  mathrmdθ\n\nNext to parameter estimation, Bayes' theorem also applies to the outer layer for model selection/comparison, updating (by the same data) model prior probabilities p(M) to posterior probabilities as \n\np(M  D) = fracp(D  M)  p(M)p(D)\n\nwhere the model evidences reappear as the essential factors, and p(D) = sum_i p(D  M_i)  p(M_i) is a normalisation factor over all models M_i (i = 1n).\n\nABC (Approximate Bayesian Computation) approximates the likelihood through model simulations [2], yielding simulated data D that are evaluated with a kernel π_ϵ(D  D),\n\nhatp(D θ M) = int  p(D  θ M)  π_ϵ(D  D)  mathrmdD\n\nwhere ϵ  0 specifies the width of the kernel. Under certain conditions, it can be shown that this approximation converges for ϵ rightarrow 0 [2], keeping only model simulations D that are \"very close\" to the original data D.\n\nThese equations show in brief, that posterior parameter and posterior model probabilities can be obtained by updating (model and parameter) priors through some data D, only requiring model simulations and a sufficiently small target ϵapprox 0. \n\nABCdeZ.jl will provide parameter samples approximating the posterior p(θ  D M) and estimations of (logarithmic) model evidences mathrmlogZ.","category":"section"},{"location":"#ABCdeZ.jl-2","page":"ABCdeZ.jl","title":"ABCdeZ.jl","text":"ABCdeZ.jl currently implements two main ABC methods:\n\nabcdesmc! (from \"abc de smc\"): A Sequential Monte Carlo algorithm. For posterior samples and model evidence estimates. Go-to method, and method of choice for multimodal or complex posterior landscapes.\nabcdemc! (from \"abc de mc\"): A simpler MCMC algorithm. For posterior samples only. \"Greedy\" algorithm that may be faster than abcdesmc! for some problems. May have difficulties with multimodal problems (use abcdesmc! instead).\n\nBoth methods make differential evolution (de) proposals for the MCMC (Markov chain Monte Carlo) steps of the algorithms; differential evolution provides a good scale and orientation to explore the (often) complex multi-dimensional parameter spaces effectively [4]. For more information on both algorithms follow the respective links above.\n\nWhy ABCdeZ.jl?\n\nIf you want to perform parameter estimation and/or model comparison, but it is difficult to compute the likelihood for your models (time-consuming and/or analytically unknow
[…]
