### 2026-09-07 e31d50f
example | minimal example extended with a fixed random seed and a small section on exact numerical reproducibility
